Reverse equals arguments for overlock AtomicMap compatibility
Overlock's atomic map wraps map values with OneShotThunk and strips them on
retrieval. Unfortunately JCTool's NonBlockingHashMap performs an `.equals` call
in various methods, and this fails when the object is not a OneShotThunk as
their equals methods (correctly) say they are not equal to a
OneShotThunk. Reversing the order means the overriden .equals method in
OneShotThunk returns true if the argument is the same value or a OneShotThunk
wrapping the same value.